Merged PR 4412: Replace old seat reallocation with atomicRearrange
Replaced seat increment and decrement with `atomicRearrange`.
`atomicRearrange` takes the `zcf` object and an array of `TransferPart` objects. Each `TransferPart` obect consists of a `fromSeat`, `toSeat`, `fromAmount` and `toAmount` and represents an allocation change between the seats.

Updated the reallocation in:
- equip
- unequip
- unequipAll
- swap
- buyItem
- buyCharacter

Since reallocation is atomic now, staged allocations have been removed. To validate the inventory, we manually construct the end state of the inventory before calling `atomicRearrange` and validate it.

Related work items: #20295