fix: permit only IST issuer and await it
In trying to use this code in a test harness, we found that the
`E(agoricNames).lookup('issuer', 'IST')` call failed because the IST
issuer had not yet been created/registered.

So rather than permitting all of `agoricNames`, pick out just
`issuer.consume.IST`, which is a _promise_ for the issuer.
Awaiting it provides the necessary synchronization.

This is not really a reduction in authority, since agoricNamesAdmin
includes agoricNames. But it makes the permit a bit more
clear/precise.