Merged PR 4423: Royalties
- Implemented royalty fee for items + characters.
- Get depositFacet of a wallet (currently hardcoded in the proposal) and pass it to the contract as privateArg
- Pass `royaltyRate` as private arg
- Updated type guard and tests
- Added check that the brand of the payment is IST (created new token (fakeIST) for tests)

When a NFT is put for sale, the royalty fee is calculated and added to the market entry.
For someone to buy this NFT, he has to give at least `askingPrice + royalty`. First we transfer the `royalty` to the `depositFacet` and then the remaining amount is given to the seller.

Related work items: #19318