Fix marketplaceAverageLevel metric (#63)
Fixed marketplaceAverageLevel metric. For items, this metric was being
calculated incorrectly. This lead to specific cases where the metric
might become a negative value after updating. This would violate the
typeguard of the metric which expects it to be greater or equal than 0
and an error would be thrown. This means the buy item flow would not
complete hence why the item would still remain in the market state.

Fixes 2 bugs:
1. #61 
2. #53

Co-authored-by: Pandelis Symeonidis <pandelis@kryha.io>