Rename .github/PULL_REQUEST_TEMPLATE/pull_request_template.md to .git… (#100)
This PR includes a fix for the PR template:

Moved the pull_request_template.md to the .github folder as having it in
a PULL_REQUEST_TEMPLATE folder requires the use of the `template` query
parameter to access the correct template. The use of the
PULL_REQUEST_TEMPLATE folder would be valid if we had multiple PR
templates. Currently we have only one so putting it directly into the
.github directory allows github to set it as the default template
automatically.

More information can be found here:
https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/creating-a-pull-request-template-for-your-repository