Frontend MakeOffer callback pattern (#104)
## Description
In this PR:

Created a new pattern for creating offer callbacks that get passed the
the wallet's `makeOffer` function. The pattern defined in
`contract-callbacks.ts` allows developers to define their callback
functions in the view components themselves. Also, having passed the
callback from the view component to the hooks you are able to overwrite
the callback and add common logic to them (see for example
`sevice/character.ts` `useBuyCharacter` hook).

```
interface MakeOfferCallback {
  accepted?: (args?: any) => void, // offer was successful
  refunded?: (args?: any) => void, // strangely seems to behave the same way as accepted
  error?: (args?: any) => void, // offer failed
  seated?: (args?: any) => void, // returned exclusively by the KREAd sell method, likely has to do with the offer being long-lived
  settled?: (args?: any) => void, // gets called when a response is received, regardless of the status
  setIsLoading?: React.Dispatch<React.SetStateAction<boolean>> 
};
```

You can now implement this interface in a view component passing it
callbacks for specific cases (error, accepted, etc). This object is
passed from the view to the hook. There common functionality can be
added to the callback by doing so:

```
{
          ...callback,
          accepted: () => {
            if (callback.accepted) callback.accepted();
            // common functionality 
          },
        },
```

This object is then passed to the service functions where it is passed
through the `formOfferResultCallback` function that generates the
complete callback that the agoric `makeOffer` function expects.

## Checklist

Make sure all items are checked before submitting the pull request.
Remove any items that are not applicable.

- [x] I have used agoric's linter on my code
(https://github.com/Agoric/agoric-sdk/discussions/8274)
- [x] I have updated the documentation to reflect the changes (if
applicable).
- [x] I have added/updated unit tests to cover the changes.
- [x] All existing tests pass.
- [x] The PR title is clear and concise.
- [x] Are there changes in the /fronted folder? Make sure `cd frontend
&& yarn build` runs successfully.;

---------

Co-authored-by: CARLOS TRIGO <carlos@kryha.io>