Fix kread bundle publishing (#112)
## Description

This PR fixes the issue with publishing the kread bundles. To fix it
`agoric published` has been replaced by the `installBundles` script in
the `agoric/contract/scripts/` folder. The script connects to the chain
using Cosmos StargateClient library and sends a transaction to the
correct endpoint with the bundle as payload. This is along the lines of
what happens under the hood of `agoric publish` as well but for some
reason that seems to be broken currently.

For funding the gas fees of the transaction, the same account is used
that funds the transactions to publish the committee bundles and sends
proposals for the kread-committee and kread.

Also, in this PR:
- Added some extra steps to `make kread-committee` (and no-build
version) that fetches the address that is to be funded from the
`AGORIC_SDK_PATH/packages/cosmic-swingset/t1/8000/` folder and funds it.
This way we no longer need to run `kread-committee`, copy the logged
address and fund it.
- Reverted back to the old folder structure where the contract code
lives in `src` instead of the versioned folder structere.

## Related Issues

Fixes #111 

## Checklist

Make sure all items are checked before submitting the pull request.
Remove any items that are not applicable.

- [x] I have used agoric's linter on my code
(https://github.com/Agoric/agoric-sdk/discussions/8274)
- [x] I have updated the documentation to reflect the changes (if
applicable).
- [x] I have added/updated unit tests to cover the changes.
- [x] All existing tests pass.